+++ 2giwon/ArchitectureStudy/data/src/main/java/com/egiwon/data/NaverRemoteDataSource.kt
@@ -1,0 +1,8 @@
+package com.egiwon.data
+
+import com.egiwon.data.model.ContentEntity
+import io.reactivex.Single
+
+interface NaverRemoteDataSource {
+    fun getRemoteContents(type: String, query: String): Single<ContentEntity>
+}
